Revert "Merge pull request #642 from BCStudentSoftwareDevTeam/Bright_Artem_manage_people"
This reverts commit 6b98e028e560309b212d81504d261d32734a73af, reversing
changes made to 60ada1ecec53be4b8241ba3c3213984376bc73b6.
