+++ app/controllers/main_routes/__init__.py
@@ -14,7 +14,6 @@ def injectGlobalData():
             'lastStaticUpdate': lastStaticUpdate}
 
 from app.controllers.main_routes import main_routes
-from app.controllers.main_routes import departmentPortal
 from app.controllers.main_routes import laborStatusForm
 from app.controllers.main_routes import laborHistory
 from app.controllers.main_routes import alterLSF
