--- app/static/css/managepositions.css
@@ -1,15 +1,0 @@
-.width-12{
-    width:12%;
-}
-
-*{
-    /* outline:solid 1px lime; */
-    margin:0;
-    padding:0;
-    box-sizing:border-box;
-}
-
-.department-header{
-    padding:1%;
-    margin-top:-20px;
-}
