+++ app/templates/main/departmentPortal.html
@@ -111,7 +111,7 @@
     </div>
 
     <div class="card-footer text-center">
-      <a href="/department/{{ department.ORG }}/{{ department.ACCOUNT }}/members" class="btn btn-primary">View Members
+      <a href="#" class="btn btn-primary">View Members
          <span class="glyphicon glyphicon-chevron-right"></span>
       </a>
     </div>
