--- app/templates/main/managepositions.html
@@ -1,90 +1,0 @@
-{% extends "base.html" %}
-
-{% block styles %}
-{{super()}}
-  <link href="/static/css/managepositions.css?u={{lastStaticUpdate}}" rel="stylesheet" type="text/css">
-{% endblock %}
-
-{% block scripts %}
-{{super()}}
-    <script type="text/javascript" src="{{url_for('static', filename='js/managePositions.js') }}?u={{lastStaticUpdate}}"></script>
-{% endblock %}
-
-{% block app_content %}
-<div class="text-white  department-header">
-    <h2>{{ department_name }}</h2>
-</div>
-
-<div class="container-fluid position-management">
-
-    <div class="row page-header">
-        <div class="form-group col-xs-4 ">
-        <select class="selectpicker"
-                name='department'
-                id='selectedDepartment'
-                data-live-search='true'
-                title="Department"
-                >
-        <option value = "" disabled selected>Search a position</option>
-        {% for p in positions %}
-            <option data-account="{{p.POSN_TITLE}}" data-org="{{p.POSN_CODE}}" data-content=
-            "<span>{{p.DEPT_NAME}}</span><small class='text-muted'> {{p.POSN_TITLE}} </small>">{{p.POSN_TITLE}}</option>
-        {% endfor %}
-        </select>
-
-  </div>
-            <h1 class="col-xs-4 text-center">Manage Positions</h1>
-        <div class="col-xs-4 text-right">
-            <button class="btn add-btn btn-primary" id="addPositionButton">
-                Add Position
-            </button>
-        </div>
-    </div>
-
-    <div class="table-responsive">
-        <table class="table table-striped table-bordered text-center" id="positionTable">
-            <thead>
-                <tr>
-                    <th class="text-center">Position (WSL)</th>
-                    <th class="text-center">Position Code</th>
-                    <th class="text-center">Status</th>
-                    <th class="width-12 text-center">Last Revision Date</th>
-                    <th class="width-12 text-center">View Position Description</th>
-                    <th class="text-center">Edit Position Description</th>
-                </tr>
-            </thead>
-            <tbody>
-                <!-- Sample Row -->
-
-                {% for position in positions %}
-                <tr>
-                    <td>{{position.POSN_TITLE}} ({{position.WLS}})</td>
-                    <td>{{position.POSN_CODE}}</td>
-                    <td>
-                        <p class="bg-success p-5">
-                            Active
-                        </p>
-                    </td>
-                    <td>01/01/2024</td>
-                    <td>
-                        <button class="btn btn-success view-btn">
-                            View
-                        </button>
-                    </td>
-                    <td>
-                        <button class="btn btn-primary request-btn">
-                            Request an Update
-                        </button>
-                    </td>
-                </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-    </div>
-    <h3 class="total">
-        Total Positions:
-        <span id="positionCount">{{ positions|length }}</span>
-    </h2>
-</div>
-
-{% endblock %}
