+++ app/templates/main/supervisorPortal.html
@@ -227,7 +227,7 @@
 </div>
 <br>
 <div style="position: relative; z-index: 1;" id="sortOptions">
-  <div class="container-fluid" style="margin-bottom: 15px; float: right;">
+  <div class="container-fluid" style="margin-bottom: -40px; float: right;">
     <label for="columnPicker">Sorting Options:</label>
     <select id="columnPicker" name="columnPicker" class="selectpicker" title="Column">
       <option value="term">Term</option>
